python-batch/requirements.txt
google-auth
google-cloud==0.34.0
google-cloud-batch==0.4.0
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.49.1
jsonschema==4.6.0
PyYaml==6.0
google-cloud-batch>=0.4.0
absl-py>=1.3.0
google-api-core>=2.10.2